requirements.txt
line-bot-sdk
flask
openai
numpy
beautifulsoup4
requests
pandas
datetime 
matplotlib
mpl_finance
gunicorn
selenium
python-dotenv